Bug correction some start-up case
In case the Mirror has been restarted during hdmi screen off (example : restart due to new config.js file or restart requested by Remote-Control), the screen stays Off and the PIR-Sensor couldn't wake it up again. It was necessary to ask Remote-Control to switch the screen On again before PIR-Sensor could work correctly again. 

To correct this issue, I propose to request as 'SCREEN_WAKEUP' on start function. As the node_helper "activateMonitor" function will check if the screen is already wakeup before to ask the hdmi output on, it has no impact on other situations.